feat: thread message delivery
Adds a notion of "senders" to the app. When joining a bubble, a user is
assigned a sender that does not overlap with any of their other existing
senders. This guarantees all per-thread comms with each user go through
a consistent contact. For now only "phone" (sms) senders are supported,
but this could evolve into email or (theoretically) other delivery
mechanisms.

When no sender is found fitting the above criteria, we purchase a new
phone # from twilio and add it to the senders DB. This is then used as
the sender for the new bubble membership, and available as a sender for
other users in the future.